requirements.txt
Flask==2.3.2
speechrecognition==3.10.0
fpdf2==2.7.4
scikit-learn==1.3.0
pytest==7.4.0
